+++ examples/plain-ruby/accounts.rb
@@ -20,3 +20,8 @@ demonstrate { account.to_h }
 demonstrate { account.upgrade }
 demonstrate { account.to_h }
 
+# Get token info for access token
+demonstrate { api.accounts.first.token_info(ENV['NYLAS_ACCESS_TOKEN']).to_h }
+
+# Getting application details
+demonstrate { api.application_details }