+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Unreleased
 * Add job status support
 * Enable account deletion
+* Add `Event` to ICS support
+* Added support for getting access token information
+* Improved support for Application Details
 
 ### 5.6.1 / 2021-12-13
 * Enabled support for adding metadata to a `NewMessage`/`Draft`